// models/SellAssetRequestModel.ts
/**
 * dfs-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2022-04-19T11:06:16Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class SellAssetRequestModel {
    'symbol': string;
    'ammountOfTokensToSell': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "symbol",
            "baseName": "symbol",
            "type": "string",
            "format": ""
        },
        {
            "name": "ammountOfTokensToSell",
            "baseName": "ammountOfTokensToSell",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return SellAssetRequestModel.attributeTypeMap;
    }

    public constructor() {
    }
}